requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.14
charset-normalizer==2.1.1
colorama==0.4.5
coverage==6.4.4
distlib==0.3.6
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
docker==6.0.0
filelock==3.8.0
flake8==5.0.4
iniconfig==1.1.1
mccabe==0.7.0
packaging==21.3
Pillow==9.2.0
pipenv==2022.9.8
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.1.3
pytz==2022.2.1
pywin32==304
requests==2.28.1
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.2
urllib3==1.26.12
virtualenv==20.16.5
virtualenv-clone==0.5.7
websocket-client==1.4.1
